Use fork join pool for map feeder.
The fork join pool provides all the desired semantics for
a work-stealing mapper tasks. The only missing functionality
is the ability to limit the size of the work queue. This is
performed with a semaphore. A permit is acquired prior to each
insertion into the queue and released upon completion of a task.
The initial number of permits determines the maximum queue length.